+++ resources/js/site.js
@@ -3,6 +3,7 @@ import collapse from '@alpinejs/collapse'
 import persist from '@alpinejs/persist'
 import trap from '@alpinejs/trap'
 import 'focus-visible'
+<<<<<<< HEAD
 var Jquery = require('jquery');
 var Slick = require("slick-carousel/slick/slick");
 
@@ -45,6 +46,8 @@ $(document).ready(function(){
 
 
 
+=======
+>>>>>>> main
 
 // Global get CSRF token function (used by forms).
 window.getToken = async () => {
@@ -57,11 +60,10 @@ window.getToken = async () => {
             this.error = 'Something went wrong. Please try again later.'
         })
 }
+
 // Call Alpine.
 window.Alpine = Alpine
 Alpine.plugin(collapse)
 Alpine.plugin(persist)
 Alpine.plugin(trap)
 Alpine.start()
-
-