+++ tailwind.config.js
@@ -7,9 +7,6 @@
 // classes. Various aspects of the config are split inmultiple files.
 //
 
-const defaultTheme = require('tailwindcss/defaultTheme')
-const plugin = require('tailwindcss/plugin')
-
 module.exports = {
   // The various configurable Tailwind configuration files.
   presets: [
@@ -17,18 +14,14 @@ module.exports = {
     require('./tailwind.config.typography.js'),
     require('./tailwind.config.peak.js'),
     require('./tailwind.config.site.js'),
-
   ],
   // Uncomment the next line to enable class based dark mode: https://peak.studio1902.nl/features/dark-mode.html.
   // darkMode: 'class',
   mode: 'jit',
   // Configure Purge CSS.
-  purge: {
-    content: [
-      './resources/views/**/*.html',
-      './resources/js/**/*.js',
-    ],
-    safelist: [],
-    layers: ['components', 'utilities'],
-  },
+  content: [
+    './resources/views/**/*.html',
+    './resources/js/**/*.js',
+  ],
+  safelist: []
 }