+++ tailwind.config.peak.js
@@ -12,8 +12,10 @@ module.exports = {
   theme: {
     extend: {
       colors: {
+        current: 'currentColor',
+        transparent: 'transparent',
         // Gray colors.
-        gray: colors.blueGray,
+        gray: colors.slate,
         // Error styling colors.
         red: colors.red,
         // Notice styling colors.
@@ -183,23 +185,9 @@ module.exports = {
 
     plugin(function({ addUtilities, theme, variants }) {
       const newUtilities = {
-        // Break words only when needed.
-        '.break-decent': {
-          wordBreak: 'break-word',
-        },
         // Fill icons that have a fill defined within their paths. For example coming from an asset container.
         '.fill-current-cascade *': {
           fill: 'currentColor',
-        },
-        // Easily switch direction on a grid.
-        '.ltr': {
-          direction: 'ltr',
-        },
-        '.rtl': {
-          direction: 'rtl',
-          '> *': {
-            direction: 'ltr',
-          },
         },
       }
       addUtilities(newUtilities)