+++ webpack.mix.js
@@ -1,7 +1,6 @@
 const mix = require('laravel-mix')
 
-
-mix.js('resources/js/site.js', 'public/js/site.js',)
+mix.js('resources/js/site.js', 'public/js/site.js')
     .sourceMaps(false)
 
 mix.postCss('resources/css/site.css', 'public/css/site.css', [
@@ -16,15 +15,15 @@ mix.options({
     cssNano: { minifyFontValues: false }
 });
 
-
-
 mix.browserSync({
     proxy: process.env.APP_URL,
     files: [
-        'resources/views/**/*.html',
-        'public/**/*.(css|js)',
+        'resources/views/**/*.html', 
+        'public/**/*.(css|js)', 
     ],
     // Option to open in non default OS browser.
     // browser: "firefox",
     notify: false
 })
+
+mix.version()